refactor: 롤로그 소개 조회/편집 lologController에 있던 소개 조회/편집 기능을 aboutController로 따로 분리하였음.
